client: show "install Docker" messages only in certain cases
Namely:
- the client is attached to a project that has Docker apps
- the client is using Science United

Also: we had two separate chunks of code for parsing all_projects_list.xml.
Combine these into one (in lib/util)
